requirements.txt
annoy==1.15.0
Django==1.11
django-annoying==0.10.4
django-bootstrap3==11.0.0
pkg-resources==0.0.0
bootstrap4==0.1.0
config==0.4.0
confusable-homoglyphs==3.2.0
dj-database-url==0.5.0
django-crispy-forms==1.7.2
django-heroku==0.3.1
django-registration==2.4.1
django-tinymce==2.8.0
djangorestframework==3.9.2
gunicorn==19.9.0
Pillow==5.4.1
psycopg2==2.7.7
python-decouple==3.1
pytz==2018.9
whitenoise==3.3.1